moved userProfile into a folder for organization
