+++ src/App.css
@@ -5,7 +5,7 @@
 }
 
 #profile-leftSideBar {
-  background-color: lightgrey;
+  background-color: rgb(240, 239, 239);
   padding: 1em;
   height: 100dvh;
   display: flex;
@@ -20,7 +20,7 @@
   padding: 0.5em;
 }
 .profile-sidebarItem:hover {
-  background-color: gray;
+  background-color: rgb(224, 224, 224);
 }
 
 .profile-Option {
@@ -35,6 +35,21 @@
   height: 200px;
   border-radius: 50%;
   overflow: hidden;
+  position: relative;
+}
+#profile-photoCover {
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.726);
+  color: white;
+  opacity: 0;
+}
+#profile-photoCover:hover {
+  opacity: 100;
 }
 #profile-userImage {
   width: 200px;
@@ -52,7 +67,7 @@
   border-radius: 5px;
 }
 .profile-svgBox:hover {
-  background-color: lightgray;
+  background-color: rgb(224, 224, 224);
 }
 
 input {
